+++ CGX_final_Project/pom.xml
@@ -217,6 +217,13 @@
 		    <artifactId>jackson-databind</artifactId>
 		    <version>2.10.1</version>
 		</dependency>
+		
+		<!-- 암호 복호화 -->
+		<dependency>
+		    <groupId>org.springframework.security</groupId>
+		    <artifactId>spring-security-core</artifactId>
+		    <version>5.2.1.RELEASE</version>
+		</dependency>
 		 
 	</dependencies>
