+++ CGX_final_Project/src/main/java/com/kh/cgx/mybatis/mybatis-config.xml
@@ -14,7 +14,7 @@ PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
 	
 	<!-- 별칭 -->
 	<typeAliases>
-
+		<typeAlias type="com.kh.cgx.entity.MemberDto" alias="member"/>
 	</typeAliases>
 	
 </configuration>
