+++ CGX_final_Project/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -34,4 +34,10 @@
 		<context:include-filter type="annotation" expression="org.springframework.stereotype.Controller"/>
 	</context:component-scan>
 	
+	<!-- 파일업로드에 필요한 도구를 등록 : CommonsMultipartResolver -->
+	<beans:bean id="multipartResolver" class="org.springframework.web.multipart.commons.CommonsMultipartResolver">
+		<beans:property name="defaultEncoding" value="UTF-8"></beans:property>
+		<beans:property name="maxUploadSize" value="1000000"></beans:property>
+	</beans:bean>
+	
 </beans:beans>
