+++ CGX_final_Project/src/main/webapp/WEB-INF/spring/root-context.xml
@@ -75,6 +75,10 @@
 		<context:exclude-filter type="annotation" expression="org.springframework.stereotype.Controller"/>
 	</context:component-scan>
 	
+
+	<!-- bcrypt password encoder 등록 -->
+	<bean id="encoder" class="org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder"></bean>
+	
 </beans>
