+++ CGX_final_Project/src/main/webapp/WEB-INF/views/user/join.jsp
@@ -1,3 +1,15 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <h1>join페이지입니다</h1>
+
+<form action="join" method="post">
+
+<label>아이디</label><input type="text" name="member_id"><br><br>
+<label>비밀번호</label><input type="text" name="member_pw"><br><br>
+<label>이름</label><input type="text" name="member_name"><br><br>
+<label>핸드폰</label><input type="text" name="member_phone"><br><br>
+<label>생일</label><input type="text" name="member_birth"><br><br>
+<label>성별</label><input type="text" name="member_sex"><br><br>
+<input type="submit" value="가입하기">
+
+</form>
