+++ CGX_final_Project/src/main/webapp/WEB-INF/views/user/login.jsp
@@ -1,3 +1,5 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <h1>login페이지입니다</h1>
+
+<input>
